Project/shodan_tool/requirements.txt
shodan==1.29.0
python-dotenv==1.0.0